Recover baby_snowworm/Tumbleweed: fix fn_8012DDB8 to 100% (if-else + peephole off), u8 param on fn_80138908 (83% partial, rlwimi matches, missing entry clrlwi)

+++ src/main/dll/baddie/Tumbleweed.c
@@ -4049,10 +4049,10 @@ void fn_801375A0(void) {
 /* EN v1.0 0x80138908  size: 24b  Bit setter at bit 6 (0x40) of obj->_b8->_58.
  * 83% — target has a leading `clrlwi r4,r4,24` that MWCC elides since
  * the rlwimi only uses bit 0 of r4. No C form found to force it. */
-void fn_80138908(int *obj, int v) {
+void fn_80138908(int *obj, u8 v) {
     u8* x = *(u8**)((char*)obj + 0xb8);
     u8 b = *(u8*)(x + 0x58);
-    *(u8*)(x + 0x58) = (u8)((b & ~0x40) | (((u8)v & 1) << 6));
+    *(u8*)(x + 0x58) = (u8)((b & ~0x40) | ((v & 1) << 6));
 }
 
 /* EN v1.0 0x80135BF0  size: 60b  titlescreen_free: if obj->_46 == 0x77d,
